Organize reinitialization better
